TL-3 pre-update
